Start, working on gameboard
I've started making the application. The cells have been started. I have a GameBoard hook that I'm attempting to work on, which utilizes the cells. Currently having issues getting it to work properly. I think that I'm taking a wrong approach in regards to making the grid for the board.