Ship Selection & Placement
The user can now click on a ship, which passes it through the ShipSelection component and into the App, which then passes the selected ship into the game board. If you have a selected ship, you can hover over a cell and it will tell you if it can be placed based off of the size of the ship. This currently only works for vertical selection, the next update will involve turning the ship horizontally by listening to an event listener.