Displaying ability to place ship (Last push did not contain the updated code)
Refactored the code to use a state for the board instead of the original array. I then worked on the updateBoard and placeShip methods to identify where a ship can be placed by color coding it as you hover over the cells you wish to place on. Next implementations will be showing the length of the ship for ability/inability to be placed and officially placing a ship on the cells.