Vertical/Horizontal Switch & Secondary Gameboard
The latest implementation refactored some of the code to include an event listener for when the character 'r' is clicked, which launches a useEffect to change the state of the direction, which is passed into the gameboard. The gameboard will then later evaluate the direction (row or col) and runs it through conditionals so that it displays the direction of the ship placement accordingly. Currently in the early stages of adding a secondary board on the page, which will represent the other player's board (i.e. where the current player's ships are and where the other play has made shots at). Currently working on the CSS on it to get it placed correctly and looking decent, the next step will be to add an onClick event for the board to place down the currently selected ship