Ship validation and placement
A ship can now be placed on the board. the placeShip has been renamed to validateShip and has been refactored to invalidate cells that are occupied. updateBoard has been refactored to include logic for placing a ship. The next step is to remove a ship from the shipSelection when it is placed.