Bugs and Ship Selection Progress
Made progress with ship selection where every time you select a ship, it is removed from the ship selection. Currently trying to add functionality to add the currently selected ship to the selection when the user attempts to select another ship while already having one selected (currently bugged). I also fixed a major bug with the ship validation where  a ship could be placed over a ship in certain situations. When all ships have been selected, a play button will replace the selection UI. However this occurs while the user still has a selected ship, which is an issue that must be resolved.